DBZ-7777 Refactor DebeziumEnginePerf to abstract base class and JMH benchmark for converting engine
